+++ doc/changes.rst
@@ -5,7 +5,7 @@ Change Log
 1.9.3 (unreleased)
 ------------------
 
-* Added :func:`desiutil.plots.plot_sky` to make all-sky plots (`PR #52`_).
+* Added new functions for creating all-sky maps (`PR #52`_) with an accompanying tutorial notebook in `doc/nb/`.
  
 .. _`PR #52`: https://github.com/desihub/desiutil/pull/52