+++ doc/changes.rst
@@ -5,10 +5,14 @@ Change Log
 1.9.3 (unreleased)
 ------------------
 
-* Added new :mod:`desiutil.sklearn` module and :class:`distutils.sklearn.GaussianMixtureModel` class to save and sample from a Gaussian mixture model.
-* Added new functions for creating all-sky maps (`PR #52`_) with an accompanying tutorial notebook in `doc/nb/`.
+* Added new :mod:`desiutil.sklearn` module and
+  :class:`distutils.sklearn.GaussianMixtureModel` class to save and
+  sample from a Gaussian mixture model.
+* Added new functions for creating all-sky maps (`PR #52`_) with an
+  accompanying tutorial notebook in `doc/nb/`.
 * Add option to :command:`fix_permissions.sh` to remove group-writeability for
   "official" data. Also, make sure that files and directories are group-readable.
+* Moved logging infrastructure from desispec.
 
 .. _`PR #52`: https://github.com/desihub/desiutil/pull/52